requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
Django==1.11.4
djangorestframework==3.6.3
kombu==4.1.0
psycopg2==2.7.3
pytz==2017.2
redis==2.10.6
vine==1.1.4
